Extract MediaTextGenerationService from MediaDocCreationService
