+++ app/services/media_doc_creation_service.rb
@@ -6,14 +6,6 @@ class MediaDocCreationService
     @attributes = attributes
   end
 
-  def generate_transcript
-    validate_medium!
-
-    @medium.file.open do |file|
-      generate_text(file.path)
-    end
-  end
-
   def save_doc(body)
     hdoc = Doc.hdoc_normalize!(
       {
@@ -29,23 +21,5 @@ class MediaDocCreationService
     doc = Doc.store_hdoc!(hdoc)
     @project.add_doc!(doc)
     doc
-  end
-
-  private
-
-  def generate_text(file_path)
-    if @medium.image?
-      ImageCaptionService.new(file_path).call
-    elsif @medium.audio?
-      AudioTranscriptionService.new(file_path).call
-    elsif @medium.video?
-      VideoTranscriptionService.new(file_path).call
-    else
-      raise ArgumentError, "Unsupported media type: #{@medium.media_type.inspect}"
-    end
-  end
-
-  def validate_medium!
-    raise ArgumentError, "Specified media has no attached file." unless @medium.file.attached?
   end
 end
